requirements.txt
responses
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.9
django-celery-beat==2.8.1
django-ipware==7.0.1
django-redis==6.0.0
django-structlog==10.0.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
kombu==5.6.2
packaging==25.0
prompt_toolkit==3.0.52
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
structlog==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
